+++ changelog.md
@@ -3,6 +3,14 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 
+## [Unreleased] - YR 
+### Added
+- views/itemnav.php
+
+### Updated
+- view/itemlist.php
+- mtce controller for pagination handling
+
 ## [Unreleased] - SC
 ### Added
 - Mtce (Maintanence Controller)